fix(scim): pass EventType to log_event, which no longer takes i32
Upstream's #7656 (fdc156b2) changed log_event's first parameter from i32 to
EventType. The three SCIM wrappers each already received a typed EventType and
immediately cast it down with `as i32`, so they broke at compile time on the
sync even though the rebase itself had no textual conflict.

Dropping the cast is the whole fix. It is also strictly better than what was
there: the downcast discarded type information at the call site only for
log_event to compare the resulting integer against hardcoded ranges, so the
three wrappers now hand the enum straight through and the compiler checks it.

rustfmt collapses the shortened calls back onto single lines, hence the -9 net.

Full suite green on the new 1.98.0 toolchain: 208 passed, 0 failed. Clippy
clean in src/api/scim/.

+++ src/api/scim/groups.rs
@@ -109,8 +109,7 @@ fn scim_group_body(group: &Group, token: &ScimToken, members: Option<&[Membershi
 }
 
 async fn log_group_event(event_type: EventType, group_uuid: &GroupId, token: &ScimToken, conn: &DbConn) {
-    log_event(event_type as i32, group_uuid, &token.org_uuid, &SCIM_ACTOR.into(), SCIM_DEVICE_TYPE, &token.ip.ip, conn)
-        .await;
+    log_event(event_type, group_uuid, &token.org_uuid, &SCIM_ACTOR.into(), SCIM_DEVICE_TYPE, &token.ip.ip, conn).await;
 }
 
 // The 400 every unresolvable member value produces. Deliberately identical for
